fix problem with clients crashing
The AddReloadListenerEvent is never fired on the client-side,
to the resources are never loaded into memory there. This makes
the client crash when it tries to resolve a breed from the Manager.

Since the data is not actually used on the client-side, we just check
for nulls up the chain and assume default values when instatiating the
Breed instead.

Additionally, with this commit I moved the code so the folder structure
aligns with the package namespaces. Intellij won't help me with
intellisense otherwise.